+++ lib/src/models/profile_circle.dart
@@ -27,7 +27,10 @@ class ProfileCircleConfiguration {
   /// Used to give padding to profile circle.
   final EdgeInsetsGeometry? padding;
 
-  /// Provides image url of user
+  /// Provides image url of user.
+  /// Or
+  /// Provides image data of user in base64
+  /// This will be determined by [isProfilePhotoInBase64].
   final String? profileImageUrl;
 
   /// Used for give bottom padding to profile circle
@@ -42,6 +45,9 @@ class ProfileCircleConfiguration {
   /// Provides callback when user long press on profile circle.
   final void Function(ChatUser)? onAvatarLongPress;
 
+  /// To check whether profile photo is in base64 or network url.
+  final bool? isProfilePhotoInBase64;
+
   const ProfileCircleConfiguration({
     this.onAvatarTap,
     this.padding,
@@ -49,5 +55,6 @@ class ProfileCircleConfiguration {
     this.bottomPadding,
     this.circleRadius,
     this.onAvatarLongPress,
+    this.isProfilePhotoInBase64,
   });
 }