+++ lib/src/widgets/chat_groupedlist_widget.dart
@@ -198,6 +198,8 @@ class _ChatGroupedListWidgetState extends State<ChatGroupedListWidget>
                   chatBubbleConfig: chatBubbleConfig?.inComingChatBubbleConfig,
                   showIndicator: widget.showTypingIndicator,
                   profilePic: profileCircleConfig?.profileImageUrl,
+                  isProfilePhotoInBase64:
+                      profileCircleConfig?.isProfilePhotoInBase64,
                 )
               : ValueListenableBuilder(
                   valueListenable: ChatViewInheritedWidget.of(context)!
@@ -209,6 +211,8 @@ class _ChatGroupedListWidgetState extends State<ChatGroupedListWidget>
                             chatBubbleConfig?.inComingChatBubbleConfig,
                         showIndicator: value,
                         profilePic: profileCircleConfig?.profileImageUrl,
+                        isProfilePhotoInBase64:
+                            profileCircleConfig?.isProfilePhotoInBase64,
                       )),
           SizedBox(
             height: (MediaQuery.of(context).size.width *